@@ -4,13 +4,16 @@
 	"description": "ZIP compress files",
 	"license": "MIT",
 	"repository": "sindresorhus/gulp-zip",
+	"funding": "https://github.com/sponsors/sindresorhus",
 	"author": {
 		"name": "Sindre Sorhus",
 		"email": "sindresorhus@gmail.com",
-		"url": "sindresorhus.com"
+		"url": "https://sindresorhus.com"
 	},
+	"type": "module",
+	"exports": "./index.js",
 	"engines": {
-		"node": ">=8"
+		"node": ">=18"
 	},
 	"scripts": {
 		"test": "xo && ava"
@@ -28,19 +31,19 @@
 		"file"
 	],
 	"dependencies": {
-		"get-stream": "^5.2.0",
-		"plugin-error": "^1.0.1",
-		"through2": "^3.0.1",
-		"vinyl": "^2.1.0",
+		"get-stream": "^8.0.1",
+		"gulp-plugin-extras": "^0.3.0",
+		"vinyl": "^3.0.0",
 		"yazl": "^2.5.1"
 	},
 	"devDependencies": {
-		"ava": "^2.3.0",
+		"ava": "^5.3.1",
 		"decompress-unzip": "^3.0.0",
+		"easy-transform-stream": "^1.0.1",
 		"gulp": "^4.0.2",
 		"vinyl-assign": "^1.2.1",
-		"vinyl-file": "^3.0.0",
-		"xo": "^0.24.0"
+		"vinyl-file": "^5.0.0",
+		"xo": "^0.56.0"
 	},
 	"peerDependencies": {
 		"gulp": ">=4"
