@@ -1,42 +1,39 @@
-'use strict';
-const path = require('path');
-const BufferConstants = require('buffer').constants;
-const Vinyl = require('vinyl');
-const PluginError = require('plugin-error');
-const through = require('through2');
-const Yazl = require('yazl');
-const getStream = require('get-stream');
+import path from 'node:path';
+import {constants as BufferConstants} from 'node:buffer';
+import Vinyl from 'vinyl';
+import Yazl from 'yazl';
+import {getStreamAsBuffer, MaxBufferError} from 'get-stream';
+import {gulpPlugin} from 'gulp-plugin-extras';
 
-module.exports = (filename, options) => {
+export default function gulpZip(filename, options) {
 	if (!filename) {
-		throw new PluginError('gulp-zip', '`filename` required');
+		throw new Error('gulp-zip: `filename` required');
 	}
 
 	options = {
 		compress: true,
 		buffer: true,
-		...options
+		...options,
 	};
 
 	let firstFile;
 	const zip = new Yazl.ZipFile();
 
-	return through.obj((file, encoding, callback) => {
+	return gulpPlugin('gulp-zip', async file => {
 		if (!firstFile) {
 			firstFile = file;
 		}
 
 		// Because Windows...
-		const pathname = file.relative.replace(/\\/g, '/');
+		const pathname = file.relative.replaceAll('\\', '/');
 
 		if (!pathname) {
-			callback();
 			return;
 		}
 
-		if (file.isNull() && file.stat && file.stat.isDirectory && file.stat.isDirectory()) {
+		if (file.isDirectory()) {
 			zip.addEmptyDirectory(pathname, {
-				mtime: options.modifiedTime || file.stat.mtime || new Date()
+				mtime: options.modifiedTime || file.stat.mtime || new Date(),
 				// Do *not* pass a mode for a directory, because it creates platform-dependent
 				// ZIP files (ZIP files created on Windows that cannot be opened on macOS).
 				// Re-enable if this PR is resolved: https://github.com/thejoshwolfe/yazl/pull/59
@@ -46,7 +43,7 @@ module.exports = (filename, options) => {
 			const stat = {
 				compress: options.compress,
 				mtime: options.modifiedTime || (file.stat ? file.stat.mtime : new Date()),
-				mode: file.stat ? file.stat.mode : null
+				mode: file.stat ? file.stat.mode : null,
 			};
 
 			if (file.isStream()) {
@@ -57,42 +54,33 @@ module.exports = (filename, options) => {
 				zip.addBuffer(file.contents, pathname, stat);
 			}
 		}
+	}, {
+		supportsAnyType: true,
+		async * onFinish() {
+			zip.end();
 
-		callback();
-	}, function (callback) {
-		if (!firstFile) {
-			callback();
-			return;
-		}
+			if (!firstFile) {
+				return;
+			}
 
-		(async () => {
 			let data;
 			if (options.buffer) {
 				try {
-					data = await getStream.buffer(zip.outputStream, {maxBuffer: BufferConstants.MAX_LENGTH});
+					data = await getStreamAsBuffer(zip.outputStream, {maxBuffer: BufferConstants.MAX_LENGTH});
 				} catch (error) {
-					if (error instanceof getStream.MaxBufferError) {
-						callback(new PluginError('gulp-zip', 'The output ZIP file is too big to store in a buffer (larger than Buffer MAX_LENGTH). To output a stream instead, set the gulp-zip buffer option to `false`.'));
-					} else {
-						callback(error);
-					}
-
-					return;
+					const error_ = error instanceof MaxBufferError ? new Error('The output ZIP file is too big to store in a buffer (larger than Buffer MAX_LENGTH). To output a stream instead, set the gulp-zip buffer option to `false`.') : error;
+					throw error_;
 				}
 			} else {
 				data = zip.outputStream;
 			}
 
-			this.push(new Vinyl({
+			yield new Vinyl({
 				cwd: firstFile.cwd,
 				base: firstFile.base,
 				path: path.join(firstFile.base, filename),
-				contents: data
-			}));
-
-			callback();
-		})();
-
-		zip.end();
+				contents: data,
+			});
+		},
 	});
-};
+}
